+++ src/components/AlbumList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/components/AlbumList/AlbumList';
