+++ src/components/NationList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/components/NationList/NationList';
